Cleanup (remove unused funcs, reorg)
