+++ swebench/harness/grading.py
@@ -3,6 +3,7 @@ from typing import Any
 from swebench.harness.constants import (
     APPLY_PATCH_FAIL,
     END_TEST_OUTPUT,
+    FAIL_ONLY_REPOS,
     FAIL_TO_FAIL,
     FAIL_TO_PASS,
     KEY_INSTANCE_ID,
@@ -19,7 +20,7 @@ from swebench.harness.constants import (
     TestStatus,
 )
 from swebench.harness.test_spec.test_spec import TestSpec
-from swebench.harness.log_parsers import MAP_REPO_TO_PARSER, get_eval_type
+from swebench.harness.log_parsers import MAP_REPO_TO_PARSER
 
 
 # MARK: Utility functions
@@ -73,7 +74,7 @@ def get_logs_eval(test_spec: TestSpec, log_fp: str) -> tuple[dict[str, str], boo
 
         # Get status map of evaluation results
         content = content.split(test_cmd)[-1]
-        return log_parser(content, test_spec), True
+        return log_parser(content), True
 
 
 def get_eval_tests_report(
@@ -264,8 +265,11 @@ def get_eval_report(
         PASS_TO_PASS: test_spec.PASS_TO_PASS,
     }
 
+    eval_type = EvalType.FAIL_ONLY if test_spec.repo in FAIL_ONLY_REPOS \
+        else EvalType.PASS_AND_FAIL
+
     report = get_eval_tests_report(
-        eval_status_map, eval_ref, eval_type=get_eval_type(test_spec)
+        eval_status_map, eval_ref, eval_type=eval_type
     )
     if get_resolution_status(report) == ResolvedStatus.FULL.value:
         report_map[instance_id]["resolved"] = True
