+++ swebench/harness/log_parsers/__init__.py
@@ -1,6 +1,5 @@
 from swebench.harness.log_parsers.javascript import MAP_REPO_TO_PARSER_JS
 from swebench.harness.log_parsers.python import MAP_REPO_TO_PARSER_PY
-from swebench.harness.log_parsers.utils import get_eval_type
 
 MAP_REPO_TO_PARSER = {
     **MAP_REPO_TO_PARSER_JS,
