+++ swebench/harness/log_parsers/javascript.py
@@ -1,13 +1,10 @@
 import re
 
-from swebench.harness.constants import (
-    TestStatus,
-)
-from swebench.harness.test_spec.test_spec import TestSpec
-from swebench.harness.log_parsers.utils import ansi_escape
+from swebench.harness.constants import TestStatus
+from swebench.harness.utils import ansi_escape
 
 
-def parse_log_calypso(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_calypso(log: str) -> dict[str, str]:
     """
     Parser for test logs generated by Calypso test suite
     """
@@ -57,7 +54,7 @@ def parse_log_calypso(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_chart_js(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_chart_js(log: str) -> dict[str, str]:
     """
     Parser for test logs generated by ChartJS test suite
     """
@@ -74,7 +71,7 @@ def parse_log_chart_js(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_marked(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_marked(log: str) -> dict[str, str]:
     """
     Parser for test logs generated by Marked test suite
     """
@@ -86,7 +83,7 @@ def parse_log_marked(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_p5js(log_content: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_p5js(log_content: str) -> dict[str, str]:
     def remove_json_blocks(log):
         filtered_lines = []
         in_json_block = False
@@ -136,7 +133,6 @@ def parse_log_p5js(log_content: str, test_spec: TestSpec) -> dict[str, str]:
             last_line_indent = line_indent
         return True
 
-    test_name_pat = re.compile(r"^(.*?)(?:\s*\(\d+(?:[A-Za-z]+)\))?$")
     log_content = ansi_escape(log_content)
     log_content = remove_json_blocks(log_content)
     log_content = remove_xml_blocks(log_content)
@@ -153,7 +149,7 @@ def parse_log_p5js(log_content: str, test_spec: TestSpec) -> dict[str, str]:
     return test_results
 
 
-def parse_log_react_pdf(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_react_pdf(log: str) -> dict[str, str]:
     """
     Parser for test logs generated by Carbon test suite
     """
