+++ swebench/harness/log_parsers/python.py
@@ -1,10 +1,9 @@
 import re
 
 from swebench.harness.constants import TestStatus
-from swebench.harness.test_spec.test_spec import TestSpec
 
 
-def parse_log_pytest(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_pytest(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with PyTest framework
 
@@ -26,7 +25,7 @@ def parse_log_pytest(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_pytest_options(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_pytest_options(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with PyTest framework with options
 
@@ -61,7 +60,7 @@ def parse_log_pytest_options(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_django(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_django(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with Django tester framework
 
@@ -141,7 +140,7 @@ def parse_log_django(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_pytest_v2(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_pytest_v2(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with PyTest framework (Later Version)
 
@@ -170,7 +169,7 @@ def parse_log_pytest_v2(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_seaborn(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_seaborn(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with seaborn testing framework
 
@@ -196,7 +195,7 @@ def parse_log_seaborn(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_sympy(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_sympy(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with Sympy framework
 
@@ -226,7 +225,7 @@ def parse_log_sympy(log: str, test_spec: TestSpec) -> dict[str, str]:
     return test_status_map
 
 
-def parse_log_matplotlib(log: str, test_spec: TestSpec) -> dict[str, str]:
+def parse_log_matplotlib(log: str) -> dict[str, str]:
     """
     Parser for test logs generated with PyTest framework
