+++ swebench/harness/utils.py
@@ -322,3 +322,10 @@ def get_modified_files(patch: str) -> list[str]:
             source_files.append(file.source_file)
     source_files = [x[2:] for x in source_files if x.startswith("a/")]
     return source_files
+
+
+def ansi_escape(text: str) -> str:
+    """
+    Remove ANSI escape sequences from text
+    """
+    return re.compile(r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])").sub("", text)
